[AudioContext.jsx] Add the functionality to shuffle the playlist
This commit adds the logic to shuffle the music playlist in the audio
context. When the shuffle button is clicked, a random music from the
playlist will be selected and played.

"[AudioContext.jsx] Add shuffle functionality to playlist

This commit implements the logic to shuffle the music playlist in the
audio context. Now, when the shuffle button is clicked, a random music
from the playlist will be selected and played. The playShuffle()
function generates a random index within the range of the playlist,
selects the corresponding song, and initiates playback."