feat: Add user info fetching to AuthContext
This change adds functionality to fetch user information and store it in
the
AuthContext. By making a GET request to the server with the user ID, the
user's
information is retrieved and stored in the `userInfo` state variable.
This will
allow convenient access to user information throughout the application.
The
`userInfo` state variable will be updated whenever the `userId` state
variable
changes. This change addresses the need for retrieving and storing user
information in the authentication context.