Refactor drop-down menu and add About and Settings pages
This commit introduces changes to the drop-down menu in the Navbar
component. It updates the condition for rendering the menu by replacing
`userInfo` with `userInfo.profile`. Additionally, two new pages, About
and Settings, are added to the application. The About page provides
information about the application, while the Settings page allows the
user to modify their profile picture and username. These changes improve
user experience and expand the functionality of the application.