Refactor app authentication and authorization logic
Simplify authentication logic in the App component to only set the user
state. Remove the unnecessary getUser() function and localStorage usage.
Use the token instead of the user ID for authentication checks. This
change ensures that the user is properly authenticated and redirects to
the stored current path after login.

In the AudioContext component, replace the userId with the token for the
authorization header when making requests. This change ensures that the
user is properly authenticated when interacting with the audio
functionality.

In the AuthContext component, replace the userId with the token for the
authorization header when making requests. Additionally, handle failed
user info fetches and clear the access_token from the localStorage when
there's an error.

In the MusicContext component, replace the userId with the token for the
authorization header when making requests.

In the CreatePlaylist component, replace the user_id field with the
token for the authorization header when making requests.

In the Login component, replace the user_id with the access_token for
storing authentication information in the localStorage. After a
successful login, redirect the user to the homepage.

In the Profile component, replace the userId with the token for the
authorization header when making requests.