Refactor MusicCard component to add navigation functionality
The commit refactors the MusicCard component in order to add navigation
functionality. The component now uses the `useNavigate` hook from
`react-router-dom` to handle navigation when a playlist is clicked. The
functionality ensures that users are directed to the appropriate page
for a given playlist or music item. This change improves the user
experience by enhancing the interactivity and navigation options within
the application.